+++ pyproject.toml
@@ -2,9 +2,7 @@
 name = "fasttrackpy"
 version = "0.1.0"
 description = "A python implementation of FastTrack"
-authors = [
-    "Josef Fruehwald <JoFrhwld@gmail.com>",
-]
+authors = ["JoFrhwld <JoFrhwld@gmail.com>"]
 license = "MIT"
 readme = "README.md"
 packages = [{include = "fasttrackpy", from="src"}]
@@ -16,34 +14,16 @@ exclude = [
 
 [tool.poetry.dependencies]
 python = ">=3.10,<3.13"
-praat-parselmouth = "^0.4.2"
-librosa = "^0.9.2"
-scipy = "^1.9.3"
-numpy = "^1.26.1"
+praat-parselmouth = "^0.4.3"
+scipy = {version = "^1.11.3", python = ">=3.10,<3.13"}
+numpy = {version = "^1.26.1", python = ">=3.10,<3.13"}
 
 [tool.poetry.group.dev.dependencies]
 jupyter = "^1.0.0"
-plotly = "^5.11.0"
-pandas = "^1.5.2"
-kaleido = "0.2.1"
 pytest = "^7.4.3"
-
-
-[tool.poetry.group.docs.dependencies]
-quartodoc = "^0.6.5"
 
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
-[tool.pytest.ini_options]
-addopts = [
-    "--import-mode=importlib",
-    "--cov-config=.coveragerc",
-    "--cov"
-]
-filterwarnings =[
-    "ignore::UserWarning"
-]
-pythonpath = "src"
-testpaths = "tests"
+pythonpath='src'
